.gitlab-ci.yml changes related to python3-flightgear packaging
Install the package whose source is in python3-flightgear inside a
temporary venv. Upgrade setuptools inside the venv to ensure it supports
PEP 660. Also remove a 'cd $CI_BUILDS_DIR' command in favor of 'cd /tmp'
(unless someone knows better) and modify the FGMeta repo URL (the
previous one had CI token vars) so as to get rid of the

  warning: redirecting to https://gitlab.com/flightgear/fgmeta.git

visible so far in the CI logs.